requirements.txt
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.7
openai==1.12.0
reportlab==4.0.4
diffusers==0.25.0
torch==2.2.0
pillow==10.0.0
imageio==2.31.0
transformers==4.37.0
accelerate==0.27.0
tqdm==4.66.1
numpy==1.24.3
safetensors==0.4.2